Lesson_VII/ matrix with linked list added

+++ exercises/sources/Lesson_V/5.9.c
@@ -24,8 +24,8 @@
 
 #define LEAP(A) ((!((A) % 4) && ((A) % 100)) || !((A) % 400))
 
-static int daytab[2][13] = {{0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31},
-			    {0, 31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31}};
+static short daytab[2][13] = {{0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31},
+			      {0, 31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31}};
 
 /* original version of K&R pag.111 */
 int day_of_year(int year, int month, int day)
@@ -61,7 +61,7 @@ int day_of_year_pointer(const int year,
   const int leap = LEAP(year);
 
   /* set a pointer 'p' to the right row */
-  int *p = &daytab[leap][1];
+  short *p = &daytab[leap][1];
   
   /* move pointer 'p' to each successive month */
   for (int i=1 ; i<month ; i++)
